oopsapidev/requirement.txt
pydantic
fastapi
uvicorn
psycopg2-binary
bcrypt
regex
pyjwt
